bugfix: do not call API when cancel on mutsig dialog   - fix display bug where MUI v5 cuts off dialogContent overflow   - move mutationsignaturetype out to common   - tslint and eslint fixes DEVSU-2955
