+++ app/common.d.ts
@@ -392,6 +392,18 @@ type SignatureVariantType = {
   variantType: 'sigv';
 } & RecordDefaults;
 
+type MutationSignatureType = {
+  associations: string | null;
+  cancerTypes: string | null;
+  features: string | null;
+  kbCategory: string | null;
+  nnls: number | null;
+  numCancerTypes: number | null;
+  pearson: number | null;
+  selected: boolean;
+  signature: string | null;
+} & RecordDefaults;
+
 type TemplateType = {
   name: string;
   headerImage: ImageType;
@@ -514,6 +526,7 @@ export {
   MsiType,
   HlaType,
   SignatureVariantType,
+  MutationSignatureType,
   MutationBurdenType,
   ImmuneType,
   MicrobialType,
