--- app/views/ReportView/components/MutationSignatures/types.d.ts
@@ -1,15 +1,0 @@
-import { RecordDefaults } from '@/common';
-
-type MutationSignatureType = {
-  associations: string | null;
-  cancerTypes: string | null;
-  features: string | null;
-  kbCategory: string | null;
-  nnls: number | null;
-  numCancerTypes: number | null;
-  pearson: number | null;
-  selected: boolean;
-  signature: string | null;
-} & RecordDefaults;
-
-export default MutationSignatureType;
